feat: Add new files and implement Bond and Gate classes
This commit adds the following changes:
- Added a new file `.crates/bin/anchor`
- Added a new file `.solitarc.js`
- Implemented the `Bond` class in `js/darkbonds/dark-bonds/accounts/Bond.ts`
- Implemented the `Gate` class in `js/darkbonds/dark-bonds/accounts/Gate.ts`

The `Bond` class represents a bond account and provides serialization and deserialization functionality for its data. It also includes methods for creating instances from arguments, deserializing from account info, fetching account info from an address, and more.

The `Gate` class represents a gate account and provides serialization and deserialization functionality for its data. It includes similar methods as the `Bond` class.

These changes were made to support the functionality of the program.

+++ programs/dark-bonds/src/bond_admin/add_lockup.rs
@@ -35,7 +35,7 @@ pub fn add_lockup(
 
     // Set these lockup values
     lockup.period = lockup_duration;
-    lockup.apy = lockup_apy;
+    lockup.apy = lockup_apy as i64;
     lockup.mature_only = mature_only;
 
     // Set purchase period
